backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
supabase==1.0.3
firebase-admin==6.2.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
alembic==1.11.1
python-dotenv==1.0.0
marshmallow==3.19.0